added back formatting but pruned problem file changed some prints to log messages added catchError for cases where deletion of locate databases fails
fixed xargs.grep args to PruneSpec to be list of strings not a list of
lists. i.e. was <new PruneSpec([parts.sublist(1)], [])> which wrapped
the list in a list. Interestingly running in checked does not catch
that since it seems to only check List<String> against List and will
not flag List<List<String>>

fixed the pruneSpec.paths arg to '-e'
moved log message about grep completion to after completion of grep.

Did that to solve issue where if I changed logging to write to stderr
I get a bad state error. The issue is, when commands are run by the
script the subcommand's stderr is mapped to the scripts.  Then when
trying to write to stderr - which was the case when I changed logger
behavior to do just that - it would fail. Kind of makes sense - if you
attach a subcommand's error stream to the scripts error stream it can
not interleave the scripts error writes into it. Anyway by moving the
log message to after the completion of the grep it did better, but the
same error popped up again from the mlocate command. This sounds
mysterious since mlocate comes first and feeds into grep, so why would
mlocate related log messages cause issues if grep should be
done. Well, it is like this:

- start grep command
- start mlocate command
- associate mlocate file descripters with grep descripters

So there are still issues. In the end I need to do no logging to
stderr while any subprocess are running if I want subprocess stderr
mapped to scripts. The purpose of trying to switch logging from simple
print to write to stderr was to allow me to get rid of grep warnings
about such and such being a directory. Rather than try to do that in
my script I decided to just use the -s grep option which suppresses
errors, even though the docs warn about its portability.